extend checking time

+++ src/xcode/ENA/ENAUITests/Home/ENAUITests_01b_Statistics.swift
@@ -123,18 +123,18 @@ class ENAUITests_01b_Statistics: CWATestCase {
 		statisticsCell.buttons[addButton].waitAndTap()
 
 		// Data selection
-		XCTAssertTrue(app.tables[AccessibilityIdentifiers.LocalStatistics.selectState].waitForExistence(timeout: .short))
+		XCTAssertTrue(app.tables[AccessibilityIdentifiers.LocalStatistics.selectState].waitForExistence(timeout: .long))
 		// Tap on some data entry. Then we should be on select district screen.
 		app.cells.element(boundBy: 1).waitAndTap()
-		XCTAssertTrue(app.tables[AccessibilityIdentifiers.LocalStatistics.selectDistrict].waitForExistence(timeout: .short))
+		XCTAssertTrue(app.tables[AccessibilityIdentifiers.LocalStatistics.selectDistrict].waitForExistence(timeout: .long))
 		// Tap on some data entry. Then we should be back on the homescreen.
 		app.cells.element(boundBy: 14).waitAndTap()
 
 		// the Local statistics card will appear.
-		XCTAssertTrue(statisticsCell.waitForExistence(timeout: .short))
+		XCTAssertTrue(statisticsCell.waitForExistence(timeout: .long))
 		app.swipeDown(velocity: .slow) // glitch
 		let localStatisticCell = statisticsCell.staticTexts[localStatisticsViewTitle]
-		XCTAssertTrue(localStatisticCell.waitForExistence(timeout: .short))
+		XCTAssertTrue(localStatisticCell.waitForExistence(timeout: .long))
 		
 		let deleteButton = statisticsCell.buttons[AccessibilityIdentifiers.General.deleteButton].firstMatch
 		XCTAssertFalse(deleteButton.isHittable)
